+++ example/tower/tower_mid.py
@@ -17,7 +17,10 @@ bp_tower_mid.block_ring_count = 30
 bp_tower_mid.even_ring_rotate = 6
 
 bp_tower_mid.render_blocks = True
+
 bp_tower_mid.render_stairs = True
+bp_tower_mid.stair_count = 12
+
 bp_tower_mid.render_window_outline = False
 
 bp_tower_mid.window_length = 20